@amriadits
Amriadi Al Masjidiy is a blockchain and cryptocurrency activist. I am currently focusing on writing on several blockchains such as Steemit.com, Mvp.gese.io, Trybe.one, Weku, Whaleshares and others.

From that, I need the right logo for my brand as a Blockchain activist. This logo can be used in Profile and Cover on the Blockchain blog. In addition, I will also use the logo on shirts and others, as a means of promotion.